+++ client/src/app/shared/shared.module.ts
@@ -83,6 +83,7 @@ import { OpenSlidesTranslateModule } from '../core/translate/openslides-translat
 import { ProjectorComponent } from './components/projector/projector.component';
 import { SlideContainerComponent } from './components/slide-container/slide-container.component';
 import { CountdownTimeComponent } from './components/contdown-time/countdown-time.component';
+import { MediaUploadContentComponent } from './components/media-upload-content/media-upload-content.component';
 
 /**
  * Share Module for all "dumb" components and pipes.
@@ -203,7 +204,8 @@ import { CountdownTimeComponent } from './components/contdown-time/countdown-tim
         SlideContainerComponent,
         OwlDateTimeModule,
         OwlNativeDateTimeModule,
-        CountdownTimeComponent
+        CountdownTimeComponent,
+        MediaUploadContentComponent
     ],
     declarations: [
         PermsDirective,
@@ -230,7 +232,8 @@ import { CountdownTimeComponent } from './components/contdown-time/countdown-tim
         MetaTextBlockComponent,
         ProjectorComponent,
         SlideContainerComponent,
-        CountdownTimeComponent
+        CountdownTimeComponent,
+        MediaUploadContentComponent
     ],
     providers: [
         { provide: DateAdapter, useClass: OpenSlidesDateAdapter },
