--- client/src/app/site/mediafiles/components/media-upload/media-upload.component.scss
@@ -1,62 +1,0 @@
-.table-container {
-    overflow: auto;
-
-    table {
-        width: 100%;
-
-        /** Title */
-        .mat-column-title {
-            min-width: 100px;
-
-            .mat-form-field {
-                width: 95%;
-            }
-        }
-
-        /** Filename */
-        .mat-column-filename {
-            min-width: 100px;
-        }
-
-        /** Information */
-        .mat-column-information {
-            min-width: 100px;
-        }
-
-        /** Hidden */
-        .mat-column-hidden {
-            min-width: 100px;
-        }
-
-        /** remove */
-        .mat-column-remove {
-            min-width: 100px;
-        }
-
-        .file-info-cell {
-            display: grid;
-            margin: 0;
-
-            span {
-                .mat-icon {
-                    font-size: 130%;
-                }
-            }
-        }
-    }
-}
-
-.selection-area {
-    margin-bottom: 1em;
-    cursor: pointer;
-}
-
-.upload-area {
-    margin-bottom: 1em;
-}
-
-.action-buttons {
-    button + button {
-        margin-left: 1em;
-    }
-}
