+++ client/src/app/site/motions/components/motion-detail/motion-detail.component.scss
@@ -177,3 +177,14 @@ span {
     font-size: 12px;
     margin-top: 4px;
 }
+
+.shortened-selector {
+    justify-content: space-between;
+    display: flex;
+    .selector {
+        width: 95%;
+    }
+    .mat-icon-button {
+        top: 20px;
+    }
+}
